Install dappmanager as dev dependency
